refactor: group LoginController dependecies in a single file"
